% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finto_api_request.R
\name{finto_api_request}
\alias{finto_api_request}
\title{Helper function to send requests to the Finto Skosmos API}
\usage{
finto_api_request(endpoint, params = list())
}
\arguments{
\item{endpoint}{API endpoint to request.}

\item{params}{List of query parameters.}
}
\value{
Parsed JSON response from the API.
}
\description{
Helper function to send requests to the Finto Skosmos API
}
\examples{
\dontrun{
finto_api_request("vocabularies")
}
}
