% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alphabetical_index_letters.R
\name{get_alphabetical_index_letters}
\alias{get_alphabetical_index_letters}
\title{Get the initial letters of the alphabetical index for labels in a specific
vocabulary from the Finto Skosmos API}
\usage{
get_alphabetical_index_letters(vocid, lang = NULL)
}
\arguments{
\item{vocid}{The vocabulary ID, e.g., "yso".}

\item{lang}{The language of labels, e.g., "en" or "fi". Optional.}
}
\value{
A tibble containing the initial letters of the alphabetical index.
}
\description{
Get the initial letters of the alphabetical index for labels in a specific
vocabulary from the Finto Skosmos API
}
\examples{
\dontrun{
result <- get_alphabetical_index_letters(vocid = "yso", lang = "fi")
print(result)
}
}
