% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annif_api.R
\name{get_annif_api_info}
\alias{get_annif_api_info}
\title{Get basic information about the Annif REST API}
\usage{
get_annif_api_info()
}
\value{
A list containing the title and version of the API.
}
\description{
This function retrieves basic information such as the title and version of
the Annif API.
}
\examples{
api_info <- get_annif_api_info()
print(api_info)
}
\author{
Akewak Jeba & Leo Lahti
}
