% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annif_project_info.R
\name{get_annif_project_info}
\alias{get_annif_project_info}
\title{Get information about a specific project from the Annif REST API}
\usage{
get_annif_project_info(project_id)
}
\arguments{
\item{project_id}{A string representing the project identifier
(e.g., "dummy-fi").}
}
\value{
A tibble containing project details, or an error message if the
project is not found.
}
\description{
This function retrieves detailed information about a specified project
using the Annif API.
}
\examples{
project_info <- get_annif_project_info(project_id = "yso-fi")
print(project_info)
}
