% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annif_vocabulary.R
\name{get_annif_vocabulary}
\alias{get_annif_vocabulary}
\title{Get a list of vocabulary from the Annif REST API}
\usage{
get_annif_vocabulary()
}
\value{
A tibble containing vocabulary details, including languages, size,
vocabulary ID,
training status, and modification time.
}
\description{
This function retrieves a list of vocabularies available in the Annif API.
}
\examples{
vocab <- get_annif_vocabulary()
print(vocab)
}
