% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concept_groups.R
\name{get_concept_groups}
\alias{get_concept_groups}
\title{Get concept groups for a given vocabulary from the Finto Skosmos API}
\usage{
get_concept_groups(vocid, lang = NULL)
}
\arguments{
\item{vocid}{The vocabulary ID, e.g., "yso".}

\item{lang}{The language of labels, e.g., "en" or "fi". Optional.}
}
\value{
A tibble containing concept groups, with fields for URI, prefLabel, and hasMembers.
}
\description{
Get concept groups for a given vocabulary from the Finto Skosmos API
}
\examples{
\dontrun{
result <- get_concept_groups(vocid = "yso", lang = "fi")
print(result)
}
}
