% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kanto.R
\name{get_kanto}
\alias{get_kanto}
\title{Process and Fetch KANTO Information using a list of author_IDs as column}
\usage{
get_kanto(data)
}
\arguments{
\item{data}{A dataframe containing an 'author_ID' column with values like "(FIN11)000069536".}
}
\value{
A tibble with \code{author_ID}, RDF data, and resolved human-readable labels.
}
\description{
This function extracts Asteri IDs from the 'author_ID' column as list,
fetches RDF data from the Finto Skosmos API using those IDs,
and returns a cleaned tibble with the retrieved metadata, profession
labels, place labels, and related-person, title, language, and
field-of-activity labels.
}
\examples{
\dontrun{
author_info <- get_kanto(fennica_subset)
}
}
