% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label_statistics.R
\name{get_label_statistics}
\alias{get_label_statistics}
\title{Get label statistics by language in a specific vocabulary from the Finto Skosmos API}
\usage{
get_label_statistics(vocid)
}
\arguments{
\item{vocid}{The vocabulary ID, e.g., "yso".}
}
\value{
A tibble containing the label statistics for each language, with columns for language, literal, property, and label count.
}
\description{
Get label statistics by language in a specific vocabulary from the Finto Skosmos API
}
\examples{
\dontrun{
result <- get_label_statistics(vocid = "yso")
print(result)
}
}
