% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_and_fetch_kanto_info.R
\name{process_and_fetch_kanto_info}
\alias{process_and_fetch_kanto_info}
\title{Extract asteriIDs, clean data, fetch RDF data from the Finto Skosmos API, and filter results}
\usage{
process_and_fetch_kanto_info(data)
}
\arguments{
\item{data}{A dataframe containing 'Code_1000' and 'Code_7001' columns.}
}
\value{
A cleaned tibble with fetched RDF data (limited to correct rows).
}
\description{
This function extracts asteriIDs, fetches RDF data, and ensures only valid rows
where \code{uri} matches the expected format (\verb{http://urn.fi/URN:NBN:fi:au:finaf:<asteriID>}).
}
\examples{
results <- process_and_fetch_kanto_info(my_data)
print(results)
}
