% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subject_suggestions.R
\name{suggest_subjects}
\alias{suggest_subjects}
\title{Suggest subjects for a given text using the Annif REST API}
\usage{
suggest_subjects(project_id, text, limit = 10, threshold = 0, language = "en")
}
\arguments{
\item{project_id}{The project identifier (e.g., "yso-en").}

\item{text}{The input text for which subjects are to be suggested.}

\item{limit}{An optional parameter to specify the maximum number of results to return. Default is 10.}

\item{threshold}{An optional parameter to specify the minimum score threshold for results. Default is 0.}

\item{language}{An optional parameter to specify the language of subject labels. Default is "en".}
}
\value{
A tibble containing suggested subjects, including their labels, URIs, and scores.
}
\description{
This function sends a request to the Annif API to suggest subjects based on the provided input text.
}
\examples{
suggestions <- suggest_subjects(project_id = "yso-en",
text = "Jean Sibelius orchestra music and composer")
print(suggestions)
}
\author{
Akewak Jeba & Leo Lahti
}
