% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yso_places.R
\name{yso_place}
\alias{yso_place}
\title{Fetch place information from YSO places (yso-paikat)}
\usage{
yso_place(x)
}
\arguments{
\item{x}{Place id, page URL, or YSO URI}
}
\value{
A one-row tibble with key fields for the place
}
\description{
Fetch place information from YSO places (yso-paikat)
}
\examples{
\dontrun{
place <- yso_place("http://www.yso.fi/onto/yso/p94257")
print(place)
}
}
