% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_profession_hierarchy.R
\name{fetch_profession_hierarchy}
\alias{fetch_profession_hierarchy}
\title{Fetch hierarchy of professions from Finto API}
\usage{
fetch_profession_hierarchy(concept_uri, vocid = "mts", depth = 3)
}
\arguments{
\item{concept_uri}{The URI of the starting profession (root).}

\item{vocid}{The vocabulary ID (default: "mts").}

\item{depth}{The depth of hierarchy to fetch (default: 3).}
}
\value{
A data.frame representing the hierarchy.
}
\description{
This function retrieves broader/narrower relationships and constructs a hierarchical structure.
}
\examples{
hierarchy_df <- fetch_profession_hierarchy("http://urn.fi/URN:NBN:fi:au:mts:m3357")
print(hierarchy_df)
}
\author{
Akewak Jeba & Leo Lahti
}
