% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_profession_info.R
\name{fetch_profession_info}
\alias{fetch_profession_info}
\title{Fetch all metadata from Finto Skosmos API}
\usage{
fetch_profession_info(concept_uri)
}
\arguments{
\item{concept_uri}{The URI of the concept to fetch.}
}
\value{
A tibble containing the full metadata.
}
\description{
This function queries the Finto API for a given concept URI and extracts all relevant metadata fields.
}
\examples{
metadata <- fetch_profession_info("http://urn.fi/URN:NBN:fi:au:mts:m3357")
print(metadata)
}
\author{
Akewak Jeba  \email{akewak84@gmail.com}
}
