% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concept_data.R
\name{get_concept_data}
\alias{get_concept_data}
\title{Get RDF data for a specific concept from the Finto Skosmos API}
\usage{
get_concept_data(uri, format = "application/json")
}
\arguments{
\item{uri}{The URI of the concept to retrieve data for.}

\item{format}{The MIME type of the serialization format (e.g., "application/rdf+xml" or "text/turtle"). Default is "application/json".}
}
\value{
A tibble containing the RDF data for the concept.
}
\description{
Get RDF data for a specific concept from the Finto Skosmos API
}
\examples{
concept_data <- get_concept_data(uri = "http://www.yso.fi/onto/afo/")
print(concept_data)
}
\author{
Akewak Jeba & Leo Lahti
}
