% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kanto.R
\name{get_kanto}
\alias{get_kanto}
\title{Process and Fetch KANTO Information using author_ID}
\usage{
get_kanto(data)
}
\arguments{
\item{data}{A dataframe containing an 'author_ID' column with values like "(FIN11)000069536".}
}
\value{
A tibble with \code{author_ID}, RDF data, and extracted profession labels.
}
\description{
This function extracts Asteri IDs from the 'author_ID' column,
fetches RDF data from the Finto Skosmos API using those IDs,
and returns a cleaned tibble with the retrieved metadata and profession labels.
}
\examples{
\dontrun{
results <- get_kanto(my_data)
}
}
\author{
Akewak Jeba  \email{akewak84@gmail.com}
}
