% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_types.R
\name{get_types}
\alias{get_types}
\title{Get information about the types (classes) of objects in all vocabularies from the Finto Skosmos API}
\usage{
get_types(lang = NULL)
}
\arguments{
\item{lang}{The language of labels to retrieve, e.g., "en" or "fi". Optional.}
}
\value{
A tibble containing the URI, label, and superclass for each type.
}
\description{
Get information about the types (classes) of objects in all vocabularies from the Finto Skosmos API
}
\examples{
result <- get_types(lang = "fi")
print(result)
}
\author{
Akewak Jeba & Leo Lahti
}
