% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vocabularies.R
\name{get_vocabularies}
\alias{get_vocabularies}
\title{Get available vocabularies from the Finto Skosmos API}
\usage{
get_vocabularies(lang = "fi")
}
\arguments{
\item{lang}{Language of labels, e.g., "en" or "fi" (default is "fi")}
}
\value{
A data frame with the vocabulary details: uri, id, and title
}
\description{
Get available vocabularies from the Finto Skosmos API
}
\examples{
result <- get_vocabularies(lang = "fi")
print(result)
}
\author{
Akewak Jeba & Leo Lahti
}
