% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vocabulary_data.R, R/get_vocabulary_types.R
\name{get_vocabulary_data}
\alias{get_vocabulary_data}
\title{Get and structure RDF data of the whole vocabulary or a specific concept from the Finto Skosmos API}
\usage{
get_vocabulary_data(
  vocid,
  format = "application/rdf+xml",
  uri = NULL,
  lang = NULL
)

get_vocabulary_data(
  vocid,
  format = "application/rdf+xml",
  uri = NULL,
  lang = NULL
)
}
\arguments{
\item{vocid}{The vocabulary ID, e.g., "yso".}

\item{format}{The MIME type of the serialization format, e.g., "text/turtle" or "application/rdf+xml". Default is "application/rdf+xml".}

\item{uri}{Optional URI of the specific concept to retrieve data for. If not provided, the whole vocabulary is returned.}

\item{lang}{Optional language code for the RDF resource, e.g., "fi" or "en".}
}
\value{
A tibble containing structured RDF data for the specified vocabulary or concept.

The RDF data in the specified format.
}
\description{
Get and structure RDF data of the whole vocabulary or a specific concept from the Finto Skosmos API

Get RDF data of a specific vocabulary or concept from the Finto Skosmos API
}
\examples{
rdf_xml_data <- get_vocabulary_data(vocid = "yso",
uri = "http://www.w3.org/2004/02/skos/core#Concept",
format = "application/rdf+xml")
cat(rdf_xml_data)
turtle_data <- get_vocabulary_data(vocid = "yso", format = "text/turtle")
cat(turtle_data)  # Print the Turtle data
result <- get_vocabulary_data(vocid = "yso", format = "text/turtle")
print(result)
}
\author{
Akewak Jeba & Leo Lahti
}
