% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_and_fetch_kanto_info.R
\name{process_and_fetch_kanto_info}
\alias{process_and_fetch_kanto_info}
\title{Process and Fetch KANTO Information}
\usage{
process_and_fetch_kanto_info(data)
}
\arguments{
\item{data}{A dataframe containing 'Code_1000' and 'Code_7001' columns.}
}
\value{
A tibble with \code{asteriID}, fetched RDF data, and profession labels.
}
\description{
This function extracts Asteri IDs from the provided dataset, fetches RDF data
from the Finto Skosmos API, and returns a cleaned tibble with the retrieved
data. The function retains the Asteri ID, the relevant API results, and
includes profession labels as separate columns.
}
\examples{
\dontrun{
results <- process_and_fetch_kanto_info(my_data)
}
}
